refactor(ipc): pr-agent 提示词抽取到 pragent-prompts 模块
run 队列执行逻辑里夹杂大段内嵌提示词（anchor marker / 排版 / 语言指示 / ask 语言后缀 /
回显去重）不合理；统一抽到 services/pragent-prompts.ts 整体维护，纯字符串构造、无 I/O：

- buildExtraInstructions：收口 EXTRA_INSTRUCTIONS 的按序拼接（语言 / anchor / 排版 / PR 上下文 / 规则）
- extraInstructionsEnvKey：EXTRA_INSTRUCTIONS 的 env key 映射
- askLanguageSuffixFor / stripAskQuestionEcho：/ask 语言后缀与输出回显去重
- languageDirectiveFor / anchorMarkerDirective / reviewLayoutDirective 为模块私有，长文本集中此处

run-queue.ts 的 executeRun 只保留「收集 prContext + 命中规则」的 I/O，再调上述函数装配 env，
瘦身约 130 行。纯结构性重构，不改运行行为；lint / typecheck / build 全绿。
